created the individual movie page
Shows the name of the movie, description, genres, spoken langues, production companies, and an image of the movie.

Also created a carousel which has data thats called by using the current movies movieID and sending that to the API to return back recommendations. Once the recommendations are gone through, it shuffles back to the first movie in the index.